requirements.txt
PyYAML==6.0.2
PyYAML==6.0.2
ultralytics==8.3.126
